src/main/java/com/rakovets/course/java/core/practice/oop_principles/BattleGround/Hero.java
package com.rakovets.course.java.core.practice.oop_principles.BattleGround;

public abstract class Hero implements Mortal {
    private String name;
    private int health;

    public Hero(String name) {
        this.name = name;
    }

    public Hero(String name, int health) {
        this.name = name;
        this.health = health;
    }

    public String getName() {
        return name;
    }

    public int getHealth() {
        return health;
    }

    public void setHealth(int health) {
        this.health = health;
    }

    public boolean isAlive() {
        return health > 0;
    }

    public void takeDamage(int damage) {
        this.health -= damage;
    }

    public abstract void attackEnemy(Enemy enemy);
}
